Move StatefulSet logic to its own module
